se corrigen warnings

+++ src/metropolis.h
@@ -2,7 +2,7 @@
 #define METROPOLIS_H
 int metropolis(int *lattice, int n, float J,float B, float *expo, float *e, int *m);
 int pick_site(int *lattice, int n);
-float flip(int *lattice, int n, int idx, float J, float B ,float *expo, float *e, int *m);
+int flip(int *lattice, int n, int idx, float J, float B ,float *expo, float *e, int *m);
 
 float energia(int *lattice, int n, float J, float B);
 int vecinos(int *lattice, int n, int idx);
